rename PhotoSetDraft to PhotoEssay
